fix(crispening): ci gate repairs for the capability wave
Fallow audit introduced-duplication gone: isParseOptions +
isSchemaCodecDataFirst hoisted into @beep/schema SchemaUtils (canonical
module-per-combinator layout, schema-topology green) and imported by
colors + file-processing; the schemaIssueToError/decodeNonNegativeInt
5-liner shared across four observability files hoisted into an
in-package internal module. redactCause overloads reordered data-last
first (the unknown-typed catch-all absorbed a lone options argument and
made the curried form unreachable). Composition test brands its capacity
literal with PosInt.make. effect-imports law --write applied (one stable
import conversion). schema-first ledger entry re-keyed after the JSDoc
insertion shifted TextSpan's inner struct to line 114.
