+++ packages/foundation/capability/observability/src/experimental/server/OtlpPacketLab.ts
@@ -7,17 +7,14 @@
 import { $ObservabilityId } from "@beep/identity/packages";
 import { LiteralKit, NonNegativeInt } from "@beep/schema";
 import { A, Str } from "@beep/utils";
-import { Clock, Context, Effect, Layer, Match, MutableRef, pipe, Result } from "effect";
+import { Clock, Context, Effect, Layer, Match, MutableRef, pipe } from "effect";
 import * as S from "effect/Schema";
 import * as OtlpSerialization from "effect/unstable/observability/OtlpSerialization";
+import { decodeNonNegativeInt } from "../../internal/decode.ts";
 import type * as O from "effect/Option";
 import type * as HttpBody from "effect/unstable/http/HttpBody";
 
 const $I = $ObservabilityId.create("experimental/server/OtlpPacketLab");
-const schemaIssueToError = (cause: S.SchemaError | S.SchemaError["issue"]): S.SchemaError =>
-  cause instanceof S.SchemaError ? cause : new S.SchemaError(cause);
-const decodeNonNegativeInt = (input: unknown) =>
-  Result.getOrThrowWith(S.decodeUnknownResult(NonNegativeInt)(input), schemaIssueToError);
 const textDecoder = new TextDecoder();
 
 /**
