+++ packages/foundation/modeling/schema/src/SchemaUtils/index.ts
@@ -9,6 +9,11 @@
  * @since 0.0.0
  * @category utilities
  */
+export * from "./isCodecDataFirst.ts";
+/**
+ * @since 0.0.0
+ * @category utilities
+ */
 export * from "./optionalKeyWithDefaults.ts";
 /**
  * @since 0.0.0
