+++ standards/schema-first.inventory.jsonc
@@ -5,7 +5,7 @@
   "entries": [
     {
       "file": "packages/foundation/capability/file-processing/src/Extraction/index.ts",
-      "symbol": "anonymous@94",
+      "symbol": "anonymous@103",
       "kind": "object-struct-schema",
       "status": "exception",
       "owner": "@beep/file-processing",
